requirements.txt
pyodbc
flask
censusname
scikit-learn
flask_cors
pytz
boto3
pyjwt
uwsgi
pandas
xlrd
XlsxWriter
apscheduler